// OregonTilth.Web/src/app/shared/models/forms/field-standard-times/field-standard-time-summary-dto.ts
import { FieldLaborActivityDto } from "../../generated/field-labor-activity-dto";
import { FieldUnitTypeDto } from "../../generated/field-unit-type-dto";
import { LaborTypeDto } from "../../generated/labor-type-dto";
import { MachineryDto } from "../../generated/machinery-dto";
import { TimeStudySummaryDto } from "../time-studies/time-studies-upsert-dto";

export class FieldStandardTimeSummaryDto {
    WorkbookID: number;
    FieldStandardTimeID: number;
    FieldLaborActivity: FieldLaborActivityDto;
    LaborType: LaborTypeDto;
    Machinery: MachineryDto;
    FieldUnitType: FieldUnitTypeDto;
    AverageMinutesPerFieldUnit: number;
    StandardTimePerUnit: number;
    TimeStudies: TimeStudySummaryDto[];
    FieldLaborActivityAndLaborTypeNameForDropdown: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }

}